Add files via upload
a tiny bit of code to gather popular phrases using Spacy. Most of the code was borrowed from an article by Analytics Vidhya. Requires a command line paramenter. I tested it on a file that contained text request mentoring services. 
I had to edit the output a bit. In future may want to remove duplicates and also do a semantic compression of the terms.